refactor: sidepannel into a directory
