+++ simplq/src/components/pages/Status/QueueDetails.jsx
@@ -4,7 +4,7 @@ import InfoIcon from '@material-ui/icons/Info';
 import styles from '../../../styles/statusPage.module.scss';
 import * as QueueService from '../../../services/queue';
 import { handleApiErrors } from '../../ErrorHandler';
-import SidePanelButton from '../../common/SidePanelButton';
+import SidePanelButton from '../../common/SidePanel/SidePanelButton';
 
 const DetailRow = ({ title, value, large }) => (
   <div className={styles['detail-row']}>
